Reran SnapATAC.ipynb with the correct snap files
Seems like previously ran SnapATAC with the old snap files, not the new
ones processed by the bulk ATAC-seq pipeline. As a result, the counts
were not the latest. This has trigerred a rerun of the data starting
from the beginning. Included metadata.tsv from the run. Rearranged some
blocks in the notebook and commented out unused ones.