Models for ChIP seq accs vs no accs
- Trained BPNet model on SOX2 ChIP seq, used the sequence representation
to train ad-hoc classifiers to predict accs vs not accs. As Anshul
pointed out, this is not ideal since the sequence rep is trained on SOX
ChIP and likely won't learn discriminative features. Abandoned this
approach.
- Trained binary factorized basset model to predict accs vs not accs for
SOX2 bound peaks (not accessible in fibroblasts). Trained with and
without initializing with bulk encode trained weights. Get moderate AUC
~0.6 not too different from SVMs.